--- custom_moduls/calender_widget/calender_events.json
@@ -1,14 +1,0 @@
-[
-    {
-        "title": "Current Project Work",
-        "start": "2024-09-10T10:00:00",
-        "end": "2024-09-10T20:00:00",
-        "resourceId": "a"
-    },
-    {
-        "title": "Event 1",
-        "start": "2024-09-10T08:30:00",
-        "end": "2024-09-10T10:30:00",
-        "resourceId": "a"
-    }
-]
