--- custom_moduls/calender_widget/calender_options.json
@@ -1,23 +1,0 @@
-{
-    "locale": "de",
-    "firstDay": "1",
-    "editable": "false",
-    "selectable": "true",
-    "headerToolbar": {
-        "left": "",
-        "center": "title",
-        "right": ""
-    },
-    "initialView": "multiMonthYear",
-    "views": {
-        "multiMonthYear": {
-            "type": "dayGridMonth",
-            "duration": {
-                "months": 18
-            },
-            "fixedWeekCount": "false"
-        }
-    },
-    "multiMonthMinWidth": 250,
-    "fixedWeekCount": "false"
-}
