+++ views/tamam_calender.py
@@ -1,3 +1,15 @@
 import streamlit as st
 
+import custom_moduls.calender_widget as cw
+
 st.write("This is the calender page.")
+
+# --- Calender ---
+# TODO: Add customizable options for the timeline
+
+cw.timeline(
+    options=cw.get_timeline_options(
+        months_ahead=18,
+        height="500px",
+    ),
+)
